ecoeaze-backend/celery/requirements.txt
celery==5.4.0
redis>=4.5.0
Pillow>=10.0.0
python-dotenv>=1.0.0
requests>=2.31.0
